% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{coef.maq}
\alias{coef.maq}
\title{Get estimates of}
\usage{
\method{coef}{maq}(object, blabl, ...)
}
\arguments{
\item{object}{The traine}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A thing
}
\description{
Gets es
}
\examples{
\donttest{
# Train a
}

}
