% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{integrated_difference}
\alias{integrated_difference}
\title{Get estimate of area the between two Qini curves with paired standard errors.}
\usage{
integrated_difference(object.lhs, object.rhs, spend)
}
\arguments{
\item{object.lhs}{A maq object \eqn{Q_a} to subtract from.}

\item{object.rhs}{A maq object \eqn{Q_b} to subtract with.}

\item{spend}{The spend level \eqn{\bar B}.}
}
\value{
An estimate of the area between the two curves along with standard errors.
}
\description{
Given two Qini curves, \eqn{Q_a} and \eqn{Q_b}, and a maximum spend \eqn{\bar B},
get an estimate of the integrated difference
\eqn{\int_{0}^{\bar B} (Q_a(B) - Q_b(B))dB}.
}
