% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{maq_scale}
\alias{maq_scale}
\title{Scale a Qini curve.}
\usage{
maq_scale(object, scale = 1)
}
\arguments{
\item{object}{A maq object.}

\item{scale}{A numeric value to scale by.}
}
\value{
A rescaled maq object.
}
\description{
Remap the policy value and budget to some problem-specific application.
This is a convenience function that is usually useful for plotting.
}
\examples{
\donttest{
# Generate some single-arm toy data.
n <- 1500
K <- 1
reward <- matrix(1 + runif(n * K), n, K)
scores <- reward + 5 * matrix(rnorm(n * K), n, K)
cost <- 1

# Fit a Qini curve.
qini <- maq(reward, cost, scores, R = 200)

if (getRversion() >= "4.1.0") { # For the pipe `|>` convenience.
# Plot the policy values as we vary the fraction treated.
qini |>
  plot(xlab = "Treated fraction")

# Plot the policy values for a maximum allocation of, for example, 500 units.
maq_scale(qini, 500) |>
  plot(xlab = "Units treated")
}
}
}
