% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{predict.maq}
\alias{predict.maq}
\title{Predict optimal treatment allocation.}
\usage{
\method{predict}{maq}(object, spend, ...)
}
\arguments{
\item{object}{A maq object.}

\item{spend}{The spend level.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A sparse matrix.
}
\description{
Gets optimal alloction matrix for a given spend level.
}
