% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{summary.maq}
\alias{summary.maq}
\title{Get}
\usage{
\method{summary}{maq}(object, ...)
}
\arguments{
\item{object}{A maq object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A thing
}
\description{
Gets estimates of
}
\examples{
\donttest{
# Train a
}

}
