% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{average_gain}
\alias{average_gain}
\title{Get estimate of gain given a spend level.}
\usage{
average_gain(object, spend)
}
\arguments{
\item{object}{A maq object.}

\item{spend}{The spend level B.}
}
\value{
An estimate of Q(B) along with standard errors.
}
\description{
Get an estimate of Q(B).
}
