% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{difference_gain}
\alias{difference_gain}
\title{Get estimate of difference in gain given a spend level with paired standard errors.}
\usage{
difference_gain(object.lhs, object.rhs, spend)
}
\arguments{
\item{object.lhs}{A maq object \eqn{Q_a} to subtract from.}

\item{object.rhs}{A maq object \eqn{Q_b} to subtract with.}

\item{spend}{The spend level B.}
}
\value{
An estimate of difference in gain along with standard errors.
}
\description{
Given two Qini curves, \eqn{Q_a} and \eqn{Q_b}, get an estimate of the difference
\eqn{Q_a(B) - Q_b(B)}, at a spend level B.
}
