% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{maq}
\alias{maq}
\title{Title}
\usage{
maq(
  reward,
  cost,
  budget,
  R = 200,
  sample.weights = NULL,
  clusters = NULL,
  tie.breaker = NULL,
  num.threads = NULL,
  seed = runif(1, 0, .Machine$integer.max)
)
}
\arguments{
\item{reward}{todo}

\item{cost}{todo}

\item{budget}{todo}

\item{R}{todo}

\item{sample.weights}{todo}

\item{clusters}{todo}

\item{tie.breaker}{todo}

\item{num.threads}{todo}

\item{seed}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
\donttest{
# Train a
}

}
