% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{maq}
\alias{maq}
\title{Fit a Multi-Action QINI.}
\usage{
maq(
  reward,
  cost,
  budget,
  reward.scores = NULL,
  R = 200,
  sample.weights = NULL,
  clusters = NULL,
  tie.breaker = NULL,
  num.threads = NULL,
  seed = runif(1, 0, .Machine$integer.max)
)
}
\arguments{
\item{reward}{A matrix of reward estimates.}

\item{cost}{A matrix of cost estimates.}

\item{budget}{The maximum spend/unit to fit the MAQ path on.}

\item{reward.scores}{A matrix of evaluation set reward score estimates.}

\item{R}{Number of bootstrap replicates for SEs. Default is 200.}

\item{sample.weights}{Weights given to an observation in estimation.
If NULL, each observation is given the same weight. Default is NULL.}

\item{clusters}{Vector of integers or factors specifying which cluster each observation corresponds to.
Default is NULL (ignored).}

\item{tie.breaker}{An optional permutation of the the integers 1 to nrow(rewards) used to
break potential ties in the optimal treatment allocation. If NULL, the ties are broken by
the lowest sample id (i.e. the sample appearing first in the data). Default is NULL.}

\item{num.threads}{Number of threads used in bootstrap replicates. By default, the number of threads
is set to the maximum hardware concurrency.}

\item{seed}{The seed of the C++ random number generator.}
}
\value{
A fit maq object.
}
\description{
Fit a Multi-Action QINI.
}
\examples{
\donttest{

if (require("grf", quietly = TRUE)) {
# Estimate treatment effects on training data.
n <- 1500
p <- 5
X <- matrix(rnorm(n * p), n, p)
W <- as.factor(sample(c("A", "B", "C"), n, replace = TRUE))
Y <- X[, 1] +  rnorm(n)
cf <- multi_arm_causal_forest(X, Y, W, W.hat = rep(1/3, 3))

# Predict CATEs and evaluation DR scores.
cate.hat <- predict(cf)$predictions[,,]
dr.hat <- get_scores(cf)[,,]

# Form cost estimates.
cost <- matrix(abs(colMeans(cate.hat)), nrow(cate.hat), ncol(cate.hat), byrow = TRUE)
cost <- n *cost / sum(cost)

# Fit MAQ using estimated CATEs and evaluate on DR scores.
max.budget <- 1
mq <- maq(cate.hat, cost, max.budget, dr.hat)

# Plot the MAQ curve.
plot(mq)

# Get an estimate of optimal reward at given spend along with standard errors.
average_gain(mq, spend = 0.1)

# Get the optimal treatment allocation matrix at a given spend.
pi.mat <- predict(mq, spend = 0.1)
}

}

}
