% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{maq}
\alias{maq}
\title{Fit a Multi-Action Qini.}
\usage{
maq(
  reward,
  cost,
  budget,
  reward.scores,
  R = 0,
  paired.inference = TRUE,
  sample.weights = NULL,
  clusters = NULL,
  tie.breaker = NULL,
  num.threads = NULL,
  seed = 42
)
}
\arguments{
\item{reward}{A matrix of reward estimates.}

\item{cost}{A matrix of cost estimates. If the costs are the same for each unit, then this can also
be a ncol(reward)-length vector.}

\item{budget}{The maximum spend per unit to fit the MAQ path on.}

\item{reward.scores}{A matrix of rewards to evaluate the MAQ on.}

\item{R}{Number of bootstrap replicates for computing standard errors. Default is 0
(only point estimates are computed).}

\item{paired.inference}{Whether to allow for paired tests with other cost curves. If TRUE (Default)
then the path of bootstrap replicates are stored in order to perform paired comparisons.
This takes memory on the order of O(RnK) and requires the comparison objects to be fit with the
same seed and R values as well as the same number of samples.}

\item{sample.weights}{Weights given to an observation in estimation.
If NULL, each observation is given the same weight. Default is NULL.}

\item{clusters}{Vector of integers or factors specifying which cluster each observation corresponds to.
Default is NULL (ignored).}

\item{tie.breaker}{An optional permutation of the the integers 1 to nrow(rewards) used to
break potential ties in the optimal treatment allocation. If NULL, the ties are broken by
the lowest sample id (i.e. the sample appearing first in the data). Default is NULL.}

\item{num.threads}{Number of threads used in bootstrap replicates. By default, the number of threads
is set to the maximum hardware concurrency.}

\item{seed}{The seed of the C++ random number generator. Default is 42.}
}
\value{
A fit maq object.
}
\description{
Fit a Multi-Action Qini.
}
\examples{
\donttest{
if (require("grf", quietly = TRUE)) {
# Fit a CATE estimator (using GRF) on a training sample.
n <- 2000
p <- 5
X <- matrix(runif(n * p), n, p)
W <- as.factor(sample(c("A", "B", "C"), n, replace = TRUE))
Y <- X[, 1] + X[, 2] * (W == "B") + X[, 3] * (W == "C") + rnorm(n)
train <- sample(1:n, n/2)
eval <- -train

tau.forest <- grf::multi_arm_causal_forest(X[train, ], Y[train], W[train])

# Predict CATEs on held out evaluation data.
tau.hat <- predict(tau.forest, X[eval, ])$predictions[,,]

# Form cost estimates - the following are a toy example.
cost.hat <- X[eval, 4:5]

# Fit an evaluation forest to compute doubly robust evaluation set scores.
eval.forest <- grf::multi_arm_causal_forest(X[eval, ], Y[eval], W[eval])
DR.scores <- grf::get_scores(eval.forest)[,,]

# Fit a MAQ with evaluation set estimates using 200 bootstrap replicates for confidence intervals.
max.budget <- 1
mq <- maq(tau.hat, cost.hat, max.budget, DR.scores, R = 200)

# Plot the MAQ curve.
plot(mq)

# Get an estimate of optimal reward at a given spend per unit along with standard errors.
average_gain(mq, spend = 0.3)

# Get the optimal treatment allocation matrix at a given spend per unit.
pi.mat <- predict(mq, spend = 0.3)

# If the treatment randomization probabilities are known, then an alternative to
# evaluation via AIPW scores is to use inverse-propensity weighting (IPW).
W.hat.true <- rep(1/3, 3)
observed.W <- match(W, levels(W))
Y.k.mat <- matrix(0, length(W), nlevels(W))
Y.k.mat[cbind(seq_along(observed.W), observed.W)] <- Y
Y.k.ipw <- sweep(Y.k.mat, 2, W.hat.true, "/")
Y.k.ipw.eval <- Y.k.ipw[eval, -1] - Y.k.ipw[eval, 1]

mq.ipw <- maq(tau.hat, cost.hat, max.budget, Y.k.ipw.eval, R = 200)

plot(mq.ipw, col = 2, add = TRUE)
legend("topleft", c("DR", "IPW", "95\% CI"), col = 1:2, lty = c(1, 1, 3))

# Compute a cost curve for only arm one.
mq.arm1 <- maq(tau.hat[, 1], cost.hat[, 1], max.budget, Y.k.ipw.eval[, 1], R = 200)
plot(mq.arm1, col = "blue", add = TRUE, ci.args = NULL)

# Get an estimate of the difference between a point on two curves along with paired standard errors.
difference_gain(mq.ipw, mq.arm1, spend = 0.3)

# Estimate a random baseline that only takes the average reward/cost into account.
mq.random <- maq(colMeans(tau.hat), colMeans(cost.hat), max.budget, Y.k.ipw.eval)
}
}

}
