% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{plot.maq}
\alias{plot.maq}
\title{Plot the gain/spend curve.}
\usage{
\method{plot}{maq}(x, ...)
}
\arguments{
\item{x}{A maq object.}

\item{...}{Additional arguments passed to plot.}
}
\description{
Plot the gain/spend curve.
}
