% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{predict.maq}
\alias{predict.maq}
\title{Predict optimal treatment allocation.}
\usage{
\method{predict}{maq}(object, spend, ...)
}
\arguments{
\item{object}{A maq object.}

\item{spend}{The spend level B.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A sparse matrix.
}
\description{
Get an estimate of the policy \eqn{\pi_B(X_i)} at a spend level B.
If there is not sufficient budget B to assign an arm to the j-th unit,
then \eqn{\pi_B(X_j)} takes on a fractional value.
}
