% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{predict.maq}
\alias{predict.maq}
\title{Predict with a}
\usage{
\method{predict}{maq}(object, spend, ...)
}
\arguments{
\item{object}{The trained}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Vector of
}
\description{
Gets optimal alloc matrix for a given spend level.
}
\examples{
\donttest{
# Train a
}

}
