% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{print.maq}
\alias{print.maq}
\title{Print a maq object.}
\usage{
\method{print}{maq}(x, ...)
}
\arguments{
\item{x}{The maq object.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print a maq object.
}
