% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{summary.maq}
\alias{summary.maq}
\title{MAQ Summary.}
\usage{
\method{summary}{maq}(object, ...)
}
\arguments{
\item{object}{A maq object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A data.frame making up the elements of the estimated Qini curve.
}
\description{
Get a data.frame with columns equal to [B, Q(B), std.err(Q(B)), i, k], where
i is the unit and k the treatment arm it is optimal to assign at a spend level B.
}
