% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maq.R
\name{summary.maq}
\alias{summary.maq}
\title{MAQ Summary.}
\usage{
\method{summary}{maq}(object, ...)
}
\arguments{
\item{object}{The maq object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A data.frame with the computed path.
}
\description{
MAQ Summary.
}
